Unnest Box.CornerStyle
