+++ BlueprintUICommonControls/Sources/Internal/UIBezierPath+Extensions.swift
@@ -5,7 +5,7 @@ extension UIBezierPath {
 
     public convenience init(
         rect: CGRect,
-        corners: Box.CornerStyle
+        corners: CornerStyle
     ) {
         switch corners {
         case .square:
